requirements.txt
# ── NOTE: pyspark KHÔNG khai báo ở đây ───────────────────────────
# bitnami/spark:3.5 đã cung cấp PySpark qua PYTHONPATH.
# Cài lại pyspark sẽ xung đột với Spark binaries của bitnami.
# Nếu muốn chạy LOCAL (ngoài Docker), tự thêm: pyspark==3.5.1

datasets==2.19.2
huggingface-hub==0.23.0
pyyaml==6.0.1
boto3==1.34.100
tqdm==4.66.4
pyarrow>=14.0.0
s3fs>=2023.6.0
polars>=0.20.0
numpy>=1.24.0
